+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Ensure the cluster role is assumed when getting endpoints subnets
+
 ## [0.2.0] - 2023-01-13
 
 ### Fixed